# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SIMS PPOB API",
    "version": "1.0.0",
    "description": "Dokumentasi API SIMS PPOB untuk Modul Membership"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": ["1. Module Membership"],
        "summary": "Login untuk mendapatkan JWT Token",
        "description": "Login untuk mendapatkan token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "janedoe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "SecurePass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "Login Sukses"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 102
                    },
                    "message": {
                      "type": "string",
                      "example": "Parameter email tidak sesuai format"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 103
                    },
                    "message": {
                      "type": "string",
                      "example": "Username atau password salah"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": ["1. Module Membership"],
        "summary": "Get Profile User",
        "description": "Mengambil data profile user berdasarkan token JWT.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 0 },
                    "message": { "type": "string", "example": "Sukses" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": { "type": "string", "example": "janedoe@example.com" },
                        "first_name": { "type": "string", "example": "Jane" },
                        "last_name": { "type": "string", "example": "Doe" },
                        "profile_image": { "type": "string", "example": "https://yoururlapi.com/profile.jpeg" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer", "example": 108 },
                    "message": { "type": "string", "example": "Token tidak valid atau kadaluwarsa" },
                    "data": { "type": "object", "nullable": true, "example": null }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}